hstore_to_json: empty hstores must return empty json objects
hstore_to_json used to return an empty string for empty hstores, but that is
not correct as an empty string is not valid json and calling row_to_json()
on rows which have empty hstores will generate invalid json for the entire
row.  The right thing to do is to return an empty json object.

Signed-off-by: Oskari Saarenmaa <os@ohmu.fi>