Truncate pg_multixact/'s contents during crash recovery.
9dc842f0832fd71eda826349a0c17ecf8ae93b84 prevented multixact
truncation during crash recovery because it wasn't guaranteed that
enough state was setup and because it wasn't deemed to be a good idea
anyway due to possible information loss. Since then due to
Hot-Standby, Streaming Rep. and PITR the amount of time a cluster can
spend in crash recovery has increased significantly to the point that
a cluster often will never come out of it. That makes not truncating
the content of pg_multixact/ not feasible anymore.

So, take care to setup enough state for multixact truncation before
crash recovery starts - easy since checkpoints contain the required
information - and move the current end-of-recovery actions to a new
TrimMultiXact() function, analogous to TrimCLOG().

At some later point this should probably done similarly to the way
clog.c is doing it, which is to just WAL log truncations, but we can't
do that for the backbranches.