Use custom macros for all C __attributes__
Previously __attribute__ was defined as an empty macro when __GNUC__ wasn't
defined as the source tree used various GCC specific attributes.  Other
compilers implement a subset of the attributes supported by GCC and we
should use them when possible, especially in the cases where they make a
functional difference like in the case of 'aligned' and 'packed'.

Replace all direct __attribute__ uses with new pg_attribute_X macros which
are defined as __attribute__(()) calls when supported by GCC, Solaris Studio
and XL C and empty in other cases.