jsonb_concat: make sure we always return a non-scalar value
jsonb_concat used to always just return the other argument if one of
arguments was considered empty.  This caused surprising behavior when
concatenating scalar values to empty lists.

Fixed this and added a test case for it.  Also added a test case for
concatenating two scalar values which currently produces an array.